models/intentionality.py
import torch
import torch.nn as nn

class IntentionalityModule(nn.Module):
    def __init__(self, hidden_dim: int, num_goals: int, num_actions: int):
        super(IntentionalityModule, self).__init__()
        self.hidden_dim = hidden_dim
        self.num_goals = num_goals
        self.num_actions = num_actions
        # ...initialize layers...

    def forward(self, state: torch.Tensor, context: torch.Tensor) -> dict:
        # ...forward pass logic...
        progress = torch.rand(state.size(0), self.num_goals)  # Non-negative
        return {
            'goals': torch.randn(state.size(0), self.num_goals, self.hidden_dim),
            'priorities': torch.randn(state.size(0), self.num_goals),
            'actions': torch.randn(state.size(0), self.num_actions),
            'progress': progress,
            'goal_progress': torch.mean(progress, dim=-1),
            'action_distributions': torch.randn(state.size(0), self.num_actions)
        }

    def update_goals(self, feedback: torch.Tensor, goals: torch.Tensor, priorities: torch.Tensor):
        # Simple placeholder update method
        new_priorities = torch.softmax(priorities + feedback, dim=-1)
        return goals, new_priorities

class GoalGenerator(nn.Module):
    def __init__(self, hidden_dim: int, num_goals: int):
        super(GoalGenerator, self).__init__()
        self.hidden_dim = hidden_dim
        self.num_goals = num_goals
        # ...initialize layers...

    def forward(self, x: torch.Tensor):
        goals = torch.randn(x.size(0), self.num_goals, self.hidden_dim)
        priorities = torch.softmax(torch.randn(x.size(0), self.num_goals), dim=-1)
        return goals, priorities
